requirements.txt
pdfminer.six
pytesseract
pdf2image
python-docx
Pillow